update: make all podcast elements as Equatable (except PodcastEpisode)
